Swapped in standard for loop instead of forEach
While the object grabbed via the getElementsByClassName is array-like, it is not a standard JavaScript array, and as such, cannot have standard array methods run on it.  Attempting a forEach on the class array results in 'Uncaught TypeError: containerEl.forEach is not a function'.  Standard for loops work as the array-like object does have index values that can address its elements.